Added Docks for Tools and for the Project Layout.
The project layout is model/view.  ProjectModel, given a molecule, sets up the
model for the molecular view.  I'm not sure if this is the correct place for
this.  I believe it could go into the libavogadro.  It's a tossup because we
may want to have our Model be the basis for interaction with the library but
the way the GL engine is setup, it takes primitives directly to render them.
This isn't a problem as we can setup our primitives to act as nodes of the
model.  Will think about this more.



git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@86 af78ad9f-9711-0410-ba57-a8eff4914ec4