ui_mainwindow.h defines Ui::MainWindow which is a private member of Avogadro::MainWindow.  QTDesigner does a crappy job with naming the layouts and embedding widgets within widgets that don't need to be that way.  There is a bit of renaming done so that we can use our FlowLayout (ripped from the tutorials).
Added icons to actions.  Images are from CrystalClear collection.

Made libavogadro const-correct as well as implementing the d-pointer technique
in glWidget and Primitives classes.

Figured out a memory leak in OpenBabel.

Updated the TreeView to match pointers rather than indexes.  This could become
a bottleneck if we have to delete a lot of primitives from a big molecule.
We'll tackle that when we cross that bridge.  Would give us something like
O(n^2) when we should really have O(n).  Although that could be wrong i was
never good with big O.

Added resource files.  Plugins now define an action which selects them.  They
can connect to the callback but it doesn't them little good.  The MainWindow
decides what tool gets called.  This just gives the plugin the ability to
define it's own tooltips / icon.  tool.h defines a default action but this
action doesn't get any information from the implemented plugin.



git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@98 af78ad9f-9711-0410-ba57-a8eff4914ec4