Remove all mentions of the USE_DISPLAY_LISTS define, always use display lists (instead of never) for the sphere and cylinder.
This fixes an issue with the way avo reused the sphere/cylinder code from Kalzium:
avo used a display list for the whole molecule, but not for the spheres and cylinders,
which caused the geometric data and the opengl traffic of each sphere/cylinder to be replicated into
the molecule display list. This caused the rendering to be unnecessarily slow and memory-consuming.



git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@159 af78ad9f-9711-0410-ba57-a8eff4914ec4