Refactored AvogadroApp -> Application to maintain naming scheme.
Added EngineListView to list the engines and allow them to be enabled/disabled
accordingly.  Engines now maintain their own PrimitiveQueue.  This will allow
us (in the future) to allow advanced users to add/remove atoms/bond/residues
accordingly from an engine.  (in the case that i only want a certain atom or
set of atoms rendered.



git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@228 af78ad9f-9711-0410-ba57-a8eff4914ec4