****Big Huge Changes****
Undo/Redo is implemented and all of the extensions and tools have been updated.
The interface to tools and extensions now requires that you return a
QUndoCommand* or 0 (NULL).  

The only important things to remember when creating QUndoCommands is to make
sure that the state after undo is exactly how it was before you performed your
action.  Worst case you need to duplicate all the Molecule data.  This is
somewhat intense but it runs quite quickly as there are no calls to NewAtom
meaning no callbacks.

There is a new class UndoSequence which is essentially a list of QUndoCommands
which are all compiled into one.  This is handy for our drawing tool.

I apologize as i haven't had time to document any of this code.



git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@303 af78ad9f-9711-0410-ba57-a8eff4914ec4