Added template for licensing.
Added navigate icon taken from the crystal clear office set.

Added "Center View" action.  There are times things are not updateing the
glwidget's geometry info.  we need to fix this so that things are required to
call it.  there is a callback from primitiveAdded / primitiveUpdate /
primitiveRemoved that should be sufficient to set a flag that the geometry data
needs to be updated.  I'll look at this more tomorrow, tomorrow, i'll love ya,
tomorrow, you're only a day away.



git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@306 af78ad9f-9711-0410-ba57-a8eff4914ec4