Added the settings dialogs to Engines.
Modified the preexisting toolSettings interface to return a 0 if there is no
settings available for the tool.  This is done to mimic the engines settings
interface and it is done in the engines settings so that we know if we should
create a new window or not.  It is better than testing against null QWidgets or
whatever the alternative would be.

Also added GLWidget::setStable(bool) and GLWidget::isStable() to see if the
GLWidget is in an unstable state.  This would occur as the result of some tool
setting it.



git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@339 af78ad9f-9711-0410-ba57-a8eff4914ec4