Committing to trunk because i have 15 min battery left and I left my AC adapter at office yesterday :/ so I can't look in the SVN book right now for how to merge, branch, switch, and whatnot.
This makes the TextRenderer work even on "old" OpenGL implementations
without GL_ARB_texture_rectangle extension.

The presence of this extension is checked for when text is first 
painted. If this extension is absent, we fall back to using square, 
power-of-two textures. Of course this is not optimal since some 
whitespace is added around each glyph, wasting a significant amount of 
texture memory, but what do you want...

anyway this is far better than 
falling back to QGLWidget::renderText(): i can't imagine how slow that 
would be on old hardware, and think of how long the lag would be at 
startup!



git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@415 af78ad9f-9711-0410-ba57-a8eff4914ec4