Added a settings dialog.
Updated the painter class to use ::begin ::end all the time.  This is good but
needs to be better.  It needs to check to make sure it's in the correct
context, if a painter is used in more than one context it should probably at
least give a warning.  Can lead to data corruption.  (check email to list)

Updated engines to use the new painter class.



git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@432 af78ad9f-9711-0410-ba57-a8eff4914ec4