Added a item model for managing engines for a GLWidget.  This will make it easier whenw e need to add / remove / edit engines.
Updated the engine interface for name and description.  

Added new configuration widget layout.  Now there is a dock widget for for
configuring an engines, this includes engine settings plus *eventually* the
primitives that are rendered by an engine.

Created the EngineTabWidget that will display the configuration for a GLWidget
(one engine at a time).  It is possible that it should be called a
GLWidgetStackedWidget but it does some weird stuff to save GUI memory.  The tab
for settings is a stack, but when we add a tab for editing primitive, this will
be a single widget for all engines.



git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@436 af78ad9f-9711-0410-ba57-a8eff4914ec4