The library now has some global functions which give runtime information about the version of the library (libVersion()), the svn revision (libSvnRevision()) and the install prefix (libPrefix()).  This should make things a big easier.  They are global functions compiled into the library at build time.
Also, removed the svn.h.in and just having cmake generate the SVN revision into
the config.h file which alreayd exists and gives information about the version
and install prefix.  Just a cleanup thing. Also, SVN revisions are different
for the library and the app now, if a commit goes into HEAD/avogadro and
HEAD/libavogadro doesnt' change then the SVN revisions will be different.  Did
this so when/if we split the two projects up we dont' have to worry about
changing this stuff.  In other words, avogadro and libavogadro get their own
config.h files generated.



git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@653 af78ad9f-9711-0410-ba57-a8eff4914ec4