woops.
the resizeGL call, which calls glViewport, must occur _before_ the 
glClear call! Otherwise only the old viewport is cleared. So there was a 
bug in the previous bugfix, and the reason why it worked nevertheless 
for Geoff is probably that the GL buffer just happened to have the 
correct initial state, but that was just luck.

Geoff: please test and tell me if it doesn't work!



git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@677 af78ad9f-9711-0410-ba57-a8eff4914ec4