Tons of updates.
Added a tabbed tool widget as requested.  This is optional.  Under settings you
can specify whether you'd like a tabbed tool widget (default) or the
traditional icon / settings layout.

Added settings.  So far this saves the following:  The current views (including
engine status, background color), the quality settings, the tabbed tool option.
Yet to come is an option to save the layout of the main window.  Such as if you
drag a dock widget somewhere.  I'm not sure exactly how this works and haven't
had time to investigate more.

The EngineFactory class now has a function to give the class name (provided by
the Qt metaobjects) of the engine they generate.  Eventually we need this to
have an engineType() function but this will come next.

You can now add new engines (of the same type as another) and you can remove
engines from your list.  If you want a default view just to do the menu and do
"View -> New View" this will make a view with the default engines.  You can
also duplicate a view this means it will copy everything except for the
primitives that the engine is maintaining.  This will also come soon.

Made it so that the docks get tabbed automatically on the left side of the
screen.  There were a lot of changes made in this and I apologize for the big
commit with little documentation on what was changed there.  Please refer to a
diff of this commit for any problems.

There is a bug fix contained in here with GLPainter not getting initialized if
the quality is set before the painter was getting initialized.

Did some code organization and cleanup in main window.



git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@827 af78ad9f-9711-0410-ba57-a8eff4914ec4