Cleaned up the openFile and loadFile functions.  loadFile is now private. MainWindow now handles all opening of files.  Please check the updated main.cpp to see how this works.  This solves a problem where if you specified a file more than once on the command line you would get multiple windows with the same file.  Also, this makes opening files look nicer, one window pops up at a time. Unfortunatly this means we have to push events through.  MainWindow::show() simply does a ::setVisible and draws the preliminary outline of the window but doesn't do a "Polish" which means that the window looks like it's stuck loading.
I made loadFile private because I believe that openFile will eventually call
loadFile but it also handles checking that the file is not already loaded in
another window.

The other change I made is that I stopped read/write of settings for the window
position.  In linux the window positioning should really be handled by the
window manager.  The other reason is that because of our settings being saved
we weren't getting any benefit.  If this needs to be implemented in mac (which
I think mac can do it's own smart window placement) then we can add it back in
for mac only.   Dimensions are still saved.


git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@1058 af78ad9f-9711-0410-ba57-a8eff4914ec4