Python scripts are now available.  This means you can write a python script (including PyQt) and it will function along with the data loaded already.  We have not yet exposed all parts of the API but this should give us the basis of where to go.
To utilize this new stuff.  Place scripts in $HOME/.avogadro/scripts

Avogadro should automatically detect them on loadup.  Also, if you modify a
script while the application is running it will reload that script before
executing it again.  See the example scripts provided.  They are installed to
$PREFIX/libavogadro/scripts but they are not automatically enabled.  Until we
get a "script manager" then we will keep it like this.

--
Donald


git-svn-id: https://avogadro.svn.sourceforge.net/svnroot/avogadro/trunk@1224 af78ad9f-9711-0410-ba57-a8eff4914ec4