GCC flags tweaked and CTestCustom.cmake added for CDash submission
I removed -Wundef from the GCC flags for now, too many warnings from the
Boost headers are generated to be useful. I also added a
CTestCustomConfig.cmake.in to ignore some warnings on Windows and for
eigen.cpp which has a massive number of warnings (this should be looked
at separately).