Use CMake to generate the include directory at build time.
This is all part of my evil plan to have CMake do all the mundane tasks
so that you have no excuse but to implement cool new features. The
Avogadro headers are now copied at CMake time into the build directory,
this is used to build against and those same headers are installed.