First network aware extension - direct PDB downloads.
This is the first network aware extension. I added QtNetwork to the
Avogadro library as I think we will want to use the network more and
more in the future. This extension is a very simple proof of concept -
it attempts to download the named PDB from the PDB site.