Relocatable Apple Mac package as part of a super project.
This is the first merge of work I have been doing to make a relocatable
Mac application bundle with all libraries included. Currently this works
if you run make install, I have been using Avogadro0.9.app as that can
be installed in parallel to the default Avogadro.app directory. Initial
testing looks pretty good, but some more work is needed to automate the
actual building of a package.