Add handling for Qt plugins including linking to the embedded Qt.
I am not sure we really want to include all the plugins, as this means
that we basically include the whole of the Qt framework too. This makes
the install size large for parts of Qt we are not using. For now I added
handling for all plugins - I think we only use the image plugins right
now. We can just more selectively install the plugins and this script
will continue to work as plugins are removed from the bundle.