Added a class to check for an updated Avogadro version.
This is the first pass at a tool to check for an updated version of
Avogadro. Feedback would be appreciated. It checks a file on the web
server, parses it and displays the release notes if the version is
greater than that installed. It will only prompt once for any given
version update, and should provide a link to our updates page.