Bring back the detached view functionality I have been missing so much!
This is me bringing back a detached GLWidget view. This is now mostly
working, with the main remaining issue being how to set the active
display to change the settings of the GLWidget. I have a few ideas,
please take a look and see if you can see any more issues with it.