Ported the bug fixes to file saving from MainWindow.
I ported the changes I made in MainWindow to fix the file saving bugs. I
also added back in some error checking for normal saves of single
molecules. Actual file saving seems to be working OK, but Avogadro does
not register the changes have been saved and so prompts the user when
closing - this needs fixing. We also need to do a lot of testing on
these functions on all three platforms to ensure we have not had any
regressions, such as the zero length file bug.