The new, improved surface extension/dialog - simpler and more intuitive.
This is the first working version of the new, improved surface dialog.
Not everything is finished yet, and it is basically a rewrite of the
orbital extension/dialog. I think that most of the logic is in there
now, need to add a little more for color mapping cubes onto each other,
showing progress dialogs for mesh calculations and fixing a few
remaining crashers. It is much simpler and allows users to accomplish
calculations using a single form.