Experimental use of the Chemical Identifier Resolver
Added support for the Chemical Identifier Resolver, found at
http://cactus.nci.nih.gov/chemical/structure. The user can type in a
structure name such as ethanol, and this service will be used to attempt
to retrieve an SDF file. Seems to work well in my initial testing. Also
switch to the Navigate tool when loading structures.

The reverse should also be possible, sending an inchi key and attempting
to resolve that to a structure name. I will look into that soon.