Added an option to reset display types to default for a view.
This was a long standing bug that caused issues for many users. I had to
add a new string to the view menu, but I think it is worth it to improve
the user experience. In addition to previous improvements to
saving/loading display type settings, the user can now reset the display
types. We should probably add a more global option to reset Avogadro to
defaults without requiring users to manually delete the configuration
file.