Removed the function level locks in the Molecule class.
The locks have been causing performance issues due to multiple
locks/unlocks in tight loops, along with some deadlocks. Migrating code
to use tryLock* gets rid of deadlocks and has improved performance in
some of the tight loops.

The Python bindings could do with some extra logic to take care of
locking, or the documentation should make it clear that scripts need to
take care of this.