Added simple color class that uses 3 floats to store color.
This is a very simple color class that uses 3 floats to store the red,
green and blue components of color. It is designed to be used with
OpenGL calls (OpenGL stores colors as floats normally), but can also be
used with other display types. It allows vectors of color objects to be
passed directly to OpenGL functions as GLfloat pointers.