Fixed some issues with file modification and saving.
Issue 2854213 should now be fixed, modification of the molecule is more
accurately reflected now. Specifically when closing a molecule the empty
window will not be marked as modified, when saving a molecule the window
will be reset to reflect it was not modified since save and unsaved new
files will prompt for a file name.