Restored the old source/library structure.
Restored the old source/library structure in Avogadro. Note that any
headers in libavogadro_HDRS are installed and part of our public API.
They cannot be removed, nor can any API incompatible changes be made.
Ideally no ABI incompatible changes can be made either, but for the
1.1/1.2 cycle this restriction is being waved.

Please be especially wary about modifying any of the headers in that
list. Look to the list in the 1.0 release for the released headers that
are part of our public API.

I also took the opportunity to clean up our CMakeLists.txt files a
little, and will continue to do so in order to make things a little
clearer.