Removed unused files, updated CMake syntax/style.
Removed a few files that are not needed anymore. Took a pass at the
CMake files to update their style, remove superfluous calls, fixed
indentation and a few other things. Builds successfully on Linux, but
it is possible that some of the platform dependent code needs more
testing.

I am not sure what some of the new compiler logic aims to do, and it
looks like most of that logic should be refactored into a support file.
Please let me know if you hit any problems.