Added initial helper script to setup development tree.
This script checks you have set up your name/email, adds a Gerrit remote
and prompts you to download the local Gerrit hook. It also adds a
prepush alias to show you what commits are about to be pushed. It also
adds a gerrit-push alias that will push the current topic to Gerrit with
the local topic branch name.

Change-Id: I8bb97400e9602d3513e87beddcd56137600f076f