Fix segfault in group indices code.
If the molecule is being modified in another thread while
calculateGroupIndices() is running, numAtoms() can be greater than
atomGroupNumber.size(). This patch triages the problem.

Ideally we should call this->lock()->lockForRead() before
calculateGroupIndices(), but the latter function is triggered by a
signal within the molecule. Since we avoid locking the mutex from
Molecule member functions, this is the best solution I can think of.

Change-Id: Id6a0f86792b7622b1ced267bd0f08e4