Optionally build against Qt Testing.
This allows Avogadro to optionally build against Qt Testing from the
ParaView project. It adds two new entries to the settings menu to record
and play back an XML test. The Qt Testing code can be obtained by,

  git clone git://paraview.org/QtTesting.git

It should be built in its own out of source tree, and you can point to
the build tree from Avogadro's build tree to use the library.

Change-Id: I7395960b912a6c1467e24fbb68db6de58dc87962