Stop recalculating group indices all of the time.
The group index is only used by the label engine, avoid calculating it
every time an atom is added/removed or the geometry is changed for
something many people never use. I am not convinced that feature has any
use outside of the label engine, and would rather see it moved there.

Change-Id: I8ce9d03798765d550bf7b74ff4fd19542fb8461b