Moved variables to d pointer, initialized them.
Some of the variables were not being initialized, such as m_groupIndex,
added them to the initializer list. Moved to the d pointer so that they
could actually move to be molecule properties in vectors in the future
(initialized lazily upon first use).

The d pointer was not being deleted, took care of that too.

Change-Id: I86b973208489be197abe73d0519d924b4ae29ed8