Clean up the SIP support, import suggested fixes.
This will also apply cleanly to the 1.0 branch. I have tested with both
master and 1.0 using SIP 4.12. Get rid of the symbol definitions - use
the ones provided by SIP instead. This means we no longer need to run
the sip binary to get the SIP version. Thanks to all the distro guys for
putting patches up pointing me in the right direction.

Change-Id: I96f8277dd63a3ad726869b1123e2238aad114c97