Extend the properties widget to handle conformers.
As we are dealing with multiple conformers, we no longer cache an OBMol
in the PropertiesModel. Instead we cache the data we need as class
variables and these are accessed by the call to data. This means that nothing
is actually calculated in the data call.

The variable m_displayConformers now controls whether we display the
data for just the current or all conformers.

Change-Id: Ie00de041b53a2bb2a56a1b42dc41a0a03b0b0918