Fix bug 3174370 with NWChem input generator
This fixes bug 3174370:
http://sourceforge.net/tracker/?func=detail&aid=3174370&group_id=165310&atid=835077

Changed so that the dedicated mp2 and ccsd modules are used, rather then
the tce engine.

mp2 jobs now run with frozen core to emulate default behaviour of
Gaussian.

Run with spherical basis sets when using Dunning correlation consistent
basis sets.

Default behaviour is now to echo the input in the output.

Change-Id: Ie3741415