Backported "Added GLWidget::renderClippedBox for unit cell rendering."
This backported code does not expose the function mentioned in the
GLWidget API, but rather implements it as a file-scoped static function
in glwidget.cpp.

The function renders a parallelepiped using OpenGL lines, and if it
intersects the viewing volume's near-plane, an appropriate line-loop
is rendered at the intersection of the near-plane and the
parallelipiped boundary. This gives the solid an apparent closed
surface to the viewer, preventing "missing corners" that are visually
disturbing.

See
http://davidlonie.blogspot.com/2011/04/gracefully-clipping-parallelepiped-in.html
for an example of what this commit does.

Change-Id: Idcba111e592d515458b8c114df5fa55af0f73f50