Changes necessary to initialize resources.
Some extra work is required to initialize the resource files from a
static plugin. Accumulate the plugins that have resource files, and then
construct a function in the static plugin header that is called once
from the PluginManager code. A few resource files needed to be renamed
to match the name of their plugin.

This now allows for an Avogadro that has all of its plugins built in
statically. These are defined at compile time using the CMake cache
variable.