Fix and expand moleculebench.cpp.
Fix issues with repeated benchmarking and added benchmarks for new
Molecule API.

The benchmark times are small and unstable, but it appears that the
atomicNum and position addAtom overload is comparable to not using the
convenience function, and the copy overload takes approximately twice as
long due to the call to the assignment operator.

The copyAtomsAndBonds functions return variable benchmark times, but are
of the same order of magnitude as just adding the atoms and bonds alone.
The mapping implementation doesn't add excessive overhead, even for
25,000 atom + 25,000 bond copies.

Change-Id: I59f0dd58cf4d59fdc77820b5d8a4399509333580