When adding a dummy atom at the center of atoms, m_selectedPrimitivesCenter was not being reset, so a subsequent add placed the atom in the wrong place.
In addition, for both the Center of Atom and Center of Mass code, no check was made whether a dummy atom was already at that location, so clicking "Add Center " was piling multiple atoms in the same location, which is almost certainly not what was intended.

This fixes those two issues.

Change-Id: Iee7d5d3d358ac1e5f7e825848c13d02edfc2aacd