When adding a dummy atom the center was not reset.
Subsequent additions placed the atom in the wrong place.

In addition, for both the Center of Atom and Center of Mass code, no check
was made whether a dummy atom was already at that location, so clicking
"Add Center " was piling multiple atoms in the same location, which is
almost certainly not what was intended.

Change-Id: Iee7d5d3d358ac1e5f7e825848c13d02edfc2aacd