Compilation fixes for Python wrapping.
See Qt bug https://bugreports.qt-project.org/browse/QTBUG-22829 for
issues introduced in Boost 1.48 that cause compilation errors in MOC.
QString is implicitly shared in fragment, and the overload was causing
compilation errors in the wrapping.

I also took a look at the fragment API, and made it more consistent with
that of the molecule. The QList is implicitly shared, and so simply
adding the missing const is enough. This then allows the Python wrapping
to compile again.

Change-Id: I678d700e81226259bed55e51800baa5601ff64e1