Compilation fixes for Python wrapping.
See Qt bug https://bugreports.qt-project.org/browse/QTBUG-22829 for
issues introduced in Boost 1.48 that cause compilation errors in MOC.
QString is implicitly shared in fragment, and the overload was causing
compilation errors in the wrapping. I also commented out the atoms and
bonds properties for now, but they should probably be added back in.

Change-Id: I678d700e81226259bed55e51800baa5601ff64e1