Quantum Espresso input generator and AvoSpg
Espresso does not use point group symmetry, instead it relies on knowing
the Bravais lattice type or the primitive lattice and the atomic basis.

Now that AvoSpg is available to all extensions, the getPrimitive
method is used to generate the primitive lattice and basis atoms.

The three crystal structure options are

1. Displayed (default)
  Take the current specification of the cell matrix and displayed atoms as
  input.  It is probably best to use a "filled" unit cell here.  I am
  going to take the assumption that Avogadro will display a conventional
  unit cell.  I believe is it not appropriate to use ibrav with this
  option as the Quantum Espresso will find the wrong symmetry.

2. Generate Primitive
  Generate the primitive lattice vectors (CELL_PARAMETERS) and basis atoms
  from the displayed unit cell.  This uses avospglib.  This option has
  been verified to produce the correct symmetry for several example
  crystal structures.

3. Manual input
  Using the displayed coordinates, the user can choose the Bravais lattice
  and the lattice parameters.

The lattice parameter can be changed irrespective of what is displayed.
Since ATOMIC_POSITIONS and CELL_PARAMETERS are in relative coordinates
for options 1 and 2 above this is very safe and allows the user to change
the lattice parameter quickly and easily without having to mess around
in the displayed molecule.

I have used the pseudopotential file list from the full archive at
http://www.quantum-espresso.org/pseudo/espresso_pp.tar.gz.

AvoSpg is now a separate library, like OpenQube.  A copy of spglib-1.2.5
is included in this commit.  (http://spglib.sourceforge.net/)

Another install of SPGLIB (libsymspg) can be used.  To force find_library
to find the correct location, the location of libsymspg can be supplied

-DAvogadro_USE_SYSTEM_SPGLIB=yes

and the patch to spglib can be supplied with

-DSGPLIB_LIBRARY=/path/to/libsymspg.so
-DSPGLIB_INCLUDE_DIR=/path/to/spglib/include

The crystallography extension has been updated to use the new AvoSpg library.

Patch 2: white space fixes

Change-Id: I4bb9d2da36f2593de74b5d058a7d74b22c5b3b2f