Add option to change cell color to crystallography extension.
Also fix custom cell color persistence.

Eventually all of this functionality needs to be moved into an engine
so that other painters (e.g. POV-ray) can correctly handle unit cells.
For now, I've moved it into the crystallography extensions because it
was the last missing bit needed to replace the old unitcellextension,
which can now be removed.

Change-Id: If413d9aa19a48c827aa4f826f6e6ad96ed309a41