AvoSpg created and some bug fixes
I had not correctly updated the number of atoms after changing the
crystal structure representation.

It appears that &electrons is sometimes required for scf.
Now &electrons and &ions will always be written to the input file.

Added the A_EXPORT to avospglib.h

AvoSpg is now a separate library, like OpenQube.  It requires that
spglib-1.2 or later (libsymspg) be installed somewhere on the system.
To force find_library to find the correct location, the location of
libsymspg can be supplied

-D SGPLIB_LIBRARY_DIR=$HOME/programs/lib

Change-Id: Ic5579c5bac4d4b0b587d1462a8e12c91f2540412