Modify the setOBMol method to fill crystal cells.
The existing code will import only the asymmetric
unit cell, which is surprising for most inorganic
crystals. (It works fine for molecular crystals.)
This patch adds a heuristic to fill the unit cell
when presented with a crystal < 4 carbon atoms.

This also removes similar code from the insert
crystal extension.

Change-Id: I07f613cebaa72bb6b51c7c6714f1e2d8533ab792