We can't perceive chains from the engine
This will break other things such as the orbitals extension. We must
perceive these chains at a point when changes can be made, ideally
adding the residues to the molecule rather than setting the entire
molecule. This fixes a segfault when loading molecules with ortbial data
when the cartoon engine is enabled (even if no biomolecule backbone is
present).

Change-Id: Ica3686cdd8f93cc88ae1638163d55af22f55c695