Added custom string options for plot axes
I have used this for my band structures. Instead of having
a number on the x axis, it is ideal for band structures to
have a symbol representing a point in reciprocal space. This
update allows me to set an axes to have strings at specific
values via the function:
setTickCustomStrings(const QList<double>& values,
                     const QStringList& strings);

If you put in a list of doubles and list of strings (they
must be matching size), then the corresponding string will
be drawn at the location given by the value in values.