Added "Reduce (Primitive and Standardize)"
There is a standard orientation for each Bravais lattice (and
perhaps space group) that a crystal needs to be in before performing
a band structure calculation on it. This is different from the
"rotate to standard orientation" already present in Avogadro. Spglib
offers the ability to do this, but it has been currently inaccessible
from the GUI. This "Reduce (Primitive and Standardize)" solves this
problem - crystals may now be oriented in their standard orientation
for band structure calculations if the user chooses this option.

This is done in light of Yaehmop additions that will be performing
band structure calculations. It is also dependent upon the newer
version of spglib (a separate pull request has been sent for that).