Added Yaehmop to Avogadro with Band Structures
The Yaehmop executable is downloaded from source forge,
and it is not able to be used until a "make install" is
performed. The yaehmop executable then gets installed to
your ${CMAKE_INSTALL_PREFIX}/bin. This needs to be a part
of your path for the Yaehmop features to run. But basically,
Yaehmop gets executed via QProcess, and then avogadro reads
what is returned. The user can execute custom input to see what
kind of output they get from Yaehmop. The user may also set a
parameters file. For now, avogadro can only send and receive
data for a band structure calculation, but it will be able to
do more in the future.