Improve usability of the animation dialog.
Prior to these changes, one was unable to (effectively) manipulate the slider in a running animation, and pressing the "play" button in a paused animation would restart the animation.
This commit, hopefully, makes the user interface more intuitive and useful, and reduces code complexity by using the more straightforward QTimer rather than QTimeLine.

Lastly, this commit makes the code correspond more closely to the user interface, by removing the `Animation::setLoopCount` function in favor of `setLoop`.
This does not remove functionality, since after all, the user is anyway only able to specify whether the animation should loop, indefinitely so, or not loop at all.
Furthermore, the need to show exactly X repetitions of the animation probably arises very rarely.
This change made changes to the `orca` and `sepctra` extensions necessary.

This commit is released to the public domain.