Added ability to show all atom images in a crystal
Added a checkbox to the crystal view options menu
(view->crystal view options...) that allows the user
to display all atom images in a crystal. It does so
by adding atoms to each engine that dislays atoms in
the needed positions.

If an atom is in a corner, images of that atom will be
placed in all the other corners.

If an atom is on an edge, images of that atom will be
placed on the other three edges it is shared with.

If an atom is on a face, images of that atom will be
placed on the opposite face.

The images are not considered to be separate atoms

This method appears to be robust and work fine when dealing
with crystal dialogs. In addition, some changes were made to
some engine classes so that the atom images may not be selected
by tools (such as the manipulate tool or draw tool). This
implementation seems to work well.