Yaehmop calc in 1-D or 2-D plus minor changes
Major change:
Spinboxes have been added to the band structure calculation
dialog and the total density of states dialog so that the user
may specify the system to be 1-D or 2-D instead of 3-D. If the user
specifies a 1-D system, periodicity will only be along the A vector
of the unit cell. If the user specifies a 2-D system, periodicity
will be along the A and B vectors of the unit cell.

Other changes:
"size_t" changed to "int" in several places where Qt containers are used
in order to reduce the "signed compared to unsigned" warnings.

numvalenceelectrons.h now have all of its valence electrons set to
exactly match that of YAeHMOP. In addition, it only goes up to the
atomic number that YAeHMOP does (90).

getNumValAndKPoints() name changed to getUserOptions()
getKPointInfo() name changed to getUserOptions()

Plotting a dashed line is cleaned up for clarity

Trapezoidal integration is now correctly divided by 2 and
the normalization factor for smoothing is multiplied by 2 to
take this into account. The DOS still integrates to the correct
numbers of electrons.

getMolecule() functions added for future use to yaehmopextension.h

Increased line widths to 2 for the plots.

Increased font size for the yaehmop output axis labels.

Also made sure that the total DOS goes to the y axis on the ends.

A few extra safety checks in yaehmopout.cpp