Fix spg change tolerance bug
When a user tries to perform spglib operations and
they fail (because the tolerance is too small), a dialog
box will appear and say "Would you like to try again with a
different tolerance"? If you say "yes", nothing happens.

This fixes that bug so that when the user says "yes", it asks
for a new tolerance, and then re-runs the operation.