Fix catastrophic bugs reading Orca orbitals.
Prior to this commit, the parser would randomly set orbital coefficients to zero for any coefficients with 3 or more characters prior to the decimal point and then silently error out for other coefficients in the affected lines.

Surprisingly this was not commonly noticeable in visualisations of several hundred orbitals, but was picked up in CAM-B3LYP calculations of ethene and WB97X calculations on butadiene.

The code is still a mess, but at least should be more robust against parsing such input files. It will be further tested in my work over the coming week or so.

Open shell orbital visualisations (copy pasted code) have not been updated as I don't have any such systems to test, and it doesn't appear to be well supported in the first place.